Removed spectrum dependency from utils
#1469
