+++ packages/jsapi-components/src/usePickerWithSelectedValues.ts
@@ -6,6 +6,7 @@ import {
   FilterConditionFactory,
 } from '@deephaven/jsapi-utils';
 import {
+  SelectionT,
   useDebouncedCallback,
   usePromiseFactory,
 } from '@deephaven/react-hooks';
@@ -13,7 +14,6 @@ import {
   COMBO_BOX_ITEM_HEIGHT,
   KeyedItem,
   SEARCH_DEBOUNCE_MS,
-  SelectionT,
   VIEWPORT_PADDING,
   VIEWPORT_SIZE,
 } from '@deephaven/utils';
