+++ packages/utils/package.json
@@ -27,8 +27,5 @@
   "sideEffects": false,
   "publishConfig": {
     "access": "public"
-  },
-  "dependencies": {
-    "@adobe/react-spectrum": "^3.29.0"
   }
 }
