+++ packages/utils/src/DataUtils.ts
@@ -1,26 +1,13 @@
-import type { Key } from 'react';
-import type { ListData } from '@adobe/react-spectrum';
-
 export const EMPTY_ARRAY = Object.freeze([]);
 
 export const EMPTY_MAP: ReadonlyMap<never, never> = new Map<never, never>();
 
 export const EMPTY_FUNCTION = () => undefined;
 
-/**
- * Subset of React Stately ListData + bulkUpdate.
- */
-export type WindowedListData<T> = Pick<
-  ListData<T>,
-  'items' | 'getItem' | 'selectedKeys' | 'setSelectedKeys' | 'update'
-> & {
-  append: (values: Iterable<T>) => void;
-  bulkUpdate: (itemMap: Map<Key, T>) => void;
-  findItem: (key: Key) => T | null;
-  insert: (index: number, values: Iterable<T>) => void;
-  remove: (keys: Iterable<Key>) => void;
-  setItems: (items: T[]) => void;
-};
+export interface KeyedItem<T> {
+  key: string;
+  item?: T;
+}
 
 /**
  * Filters out any null or undefined values from an array.
