+++ packages/utils/src/index.ts
@@ -6,6 +6,7 @@ export * from './ClipboardUtils';
 export * from './DataUtils';
 export { default as DbNameValidator } from './DbNameValidator';
 export { default as DOMUtils } from './DOMUtils';
+export * from './DOMUtils';
 export * from './EventTargetShimUtils';
 export * from './hasRequiredString';
 export * from './isNonEmptyString';
@@ -16,8 +17,6 @@ export * from './Asserts';
 export * from './ErrorUtils';
 export { default as RangeUtils, generateRange } from './RangeUtils';
 export type { Range } from './RangeUtils';
-export * from './SelectionUtils';
-export * from './SpectrumUtils';
 export { default as TextUtils } from './TextUtils';
 export { default as TimeoutError } from './TimeoutError';
 export { default as TimeUtils } from './TimeUtils';
